Moving MedalFactory from models.py to factories.py.
Added a conditional to return the created medal only if it has been saved to DB
